Add distributed Redis integration tests using miniredis (closes #60)
Refactor pkg/tokens/integration to follow the shared-suite pattern from
pkg/storage, and add Redis-specific distributed behavior tests.

- Replace monolithic integration_test.go with three focused files:
  - integration_suite_test.go: Ginkgo bootstrap + ServiceFactory type +
    RunTokenServiceIntegrationTests() shared function with all 10 behavioral
    specs; secondary services created via factory (not hardcoded DiskKeyStore);
    store2.Cleanup() replaced with svc2.CleanupExpiredTokens() (public API)
  - disk_memory_test.go: DiskKeyStore + MemoryRefreshStore factory, calls
    the shared suite
  - redis_test.go: RedisKeyStore + RedisRefreshStore factory (isolated
    miniredis per factory call), calls the shared suite; adds 2 distributed
    tests validating cross-instance revocation and cross-instance key loading

- Update run-ci-locally.sh and CI.yml:
  - Rename integration step to reflect both backends
  - Add --randomize-all for better isolation
  - Increase --timeout from 120s to 180s

Result: 22 integration specs (10 disk+memory + 10 Redis + 2 distributed),
all passing race-detection clean. No external Redis required (miniredis).
